- Improved code structure, decoding, and file handling for Transactions
